+++ .golangci.yaml
@@ -36,3 +36,6 @@ linters-settings:
   #  local-prefixes: github.com/databricks/cli
 issues:
   exclude-dirs-use-default: false  # recommended by docs https://golangci-lint.run/usage/false-positives/
+run:
+  build-tags:
+    - integration
