+++ integration/cmd/jobs/jobs_test.go
@@ -18,7 +18,7 @@ func TestAccCreateJob(t *testing.T) {
 	if env != "azure" {
 		t.Skipf("Not running test on cloud %s", env)
 	}
-	stdout, stderr := testcli.RequireSuccessfulRun(t, "jobs", "create", "--json", "@testjsons/create_job_without_workers.json", "--log-level=debug")
+	stdout, stderr := testcli.RequireSuccessfulRun(t, "jobs", "create", "--json", "@testdata/create_job_without_workers.json", "--log-level=debug")
 	assert.Empty(t, stderr.String())
 	var output map[string]int
 	err := json.Unmarshal(stdout.Bytes(), &output)
