+++ integration/cmd/storage_credentials/storage_credentials_test.go
