+++ integration/cmd/unknown_command_test.go
@@ -1,4 +1,6 @@
-package internal
+//go:build integration
+
+package cmd_integration
 
 import (
 	"testing"
