+++ integration/cmd/workspace/testdata/import_dir/a/b/c/file-b
