+++ integration/libs/filer/filer_test.go
@@ -1,4 +1,6 @@
-package internal
+//go:build integration
+
+package filer_integration
 
 import (
 	"bytes"
