+++ integration/libs/locker/locker_test.go
@@ -1,4 +1,6 @@
-package internal
+//go:build integration
+
+package locker_integration
 
 import (
 	"context"
